pyproject.toml
[tool.poetry]
name = "todo_list_flask_brython"
version = "0.1.0"
description = ""
authors = ["dunossauro <mendesxeduardo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
sqlalchemy = "^1.3.18"
flask-marshmallow = "^0.13.0"
marshmallow-sqlalchemy = "^0.23.1"
flask-migrate = "^2.5.3"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-sugar = "^0.9.4"
pytest-html = "^2.1.1"
ipdb = "^0.13.3"
pytest-mock = "^3.2.0"
selenium = "^3.141.0"
factory_boy = "^2.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
